+++ server/www/oauth/connections.js
@@ -1,7 +1,8 @@
 const API = require('../../utils/api');
 const Providers = require('./providers');
+const GoogleAPIs = require('../../utils/oauthConnections').GoogleAPIs;
+const FacebookMarketing = require('../../utils/oauthConnections').FacebookMarketing;
 const fetch = require('node-fetch');
-const constants = require('../../utils/constants');
 
 exports.list = class extends API {
 
@@ -145,356 +146,5 @@ exports.test = class extends API {
 		this.assert(connection, 'Invalid provider type!');
 
 		return await connection.test();
-	}
-}
-
-exports.gaReports = class extends API {
-
-	async gaReports({id, viewId, startDate, endDate, metrics, dimensions} = {}) {
-
-		const [oAuthProvider] = await this.mysql.query(
-			`SELECT
-				c.id connection_id,
-				c.access_token,
-				c.refresh_token,
-				p.*
-			FROM
-				tb_oauth_connections c
-			JOIN
-				tb_oauth_providers p USING (provider_id)
-			WHERE
-				c.id = ? 
-				AND c.user_id = ?
-				AND c.status = 1`,
-			[id, this.user.user_id]
-		);
-
-		this.assert(oAuthProvider, 'Invalid connection id');
-
-        let connection;
-
-        if(oAuthProvider.type == 'Google') {
-
-            connection =  new GoogleAPIs(this, oAuthProvider);
-        }
-
-        this.assert(connection, 'Invalid provider type!');
-
-        const access_token = await connection.test();
-
-		const response = await this.fetch(viewId, startDate, endDate, metrics, dimensions, access_token);
-
-		const processedData = [];
-
-		for(const row of response.data.rows) {
-
-			const rowObj = {};
-
-			for(const [i, dimension] of row.dimensions.entries()) {
-
-				rowObj[response.columnHeader.dimensions[i].replace('ga:', '')] = dimension;
-			}
-
-			for(const [i, metric] of row.metrics[0].values.entries()) {
-
-				rowObj[response.columnHeader.metricHeader.metricHeaderEntries[i].name.replace('ga:', '')] = metric;
-			}
-
-			processedData.push(rowObj);
-		}
-
-		if(!this.account.settings.has("load_saved_connection")) {
-
-			return {
-				status: true,
-				data: "save connection missing"
-			}
-		}
-		else {
-
-			const database = await this.initializeDb();
-
-			const insertResponse = await this.mysql.query(
-				`INSERT INTO ${database}.${constants.saveQueryResultTable} (type, user_id, data) VALUES ('ga', ?, ?)`,
-				[this.user.user_id, JSON.stringify(processedData)],
-				this.account.settings.get("load_saved_connection")
-			);
-
-			return {
-				status: true,
-				data: {
-					message: `Data saved in the table ${constants.saveQueryResultTable}`,
-					response: insertResponse
-				}
-			}
-		}
-
-	}
-
-	async fetch(viewId, startDate, endDate, metrics, dimensions, access_token) {
-
-		const gaMetrics = [], gaDimensions = [];
-
-		for(const metric of typeof metrics == 'string' ? [metrics] : metrics) {
-
-			gaMetrics.push({
-				"expression": `ga:${metric}`
-			});
-		}
-
-		for(const dimension of typeof dimensions == 'string' ? [dimensions] : dimensions) {
-
-			gaDimensions.push({
-				"name": `ga:${dimension}`
-			});
-		}
-
-		const
-			options = {
-				method: 'POST',
-				body: JSON.stringify({
-					'reportRequests': [
-						{
-							'viewId': viewId,
-							'dateRanges': [{'startDate': startDate, 'endDate': endDate}],
-							'metrics': gaMetrics,
-							"dimensions": gaDimensions
-						}
-					]
-				}),
-				headers: {
-					'Content-Type': 'application/json',
-					'Authorization': `Bearer ${access_token}`
-				},
-			};
-
-		let response = await fetch('https://analyticsreporting.googleapis.com/v4/reports:batchGet', options);
-
-		response = await response.json();
-
-		return response.reports[0];
-	}
-
-	async initializeDb() {
-
-		let
-			conn = this.account.settings.get("load_saved_connection"),
-			savedDatabase = this.account.settings.get("load_saved_database"),
-			db = await this.mysql.query("show databases", [], conn);
-
-		[db] = db.filter(x => x === (savedDatabase || constants.saveQueryResultDb));
-
-		if (!db) {
-
-			await this.mysql.query(
-				`CREATE DATABASE IF NOT EXISTS ${savedDatabase || constants.saveQueryResultDb}`,
-				[],
-				conn
-			);
-		}
-
-		await this.mysql.query(`
-			CREATE TABLE IF NOT EXISTS ??.?? (
-			  \`id\` int(11) unsigned NOT NULL AUTO_INCREMENT,
-			  \`type\` varchar(20) DEFAULT NULL,
-			  \`user_id\` int(11) DEFAULT NULL,
-			  \`data\` longblob,
-			  \`created_at\` timestamp NULL DEFAULT CURRENT_TIMESTAMP,
-			  \`updated_at\` timestamp NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
-			  PRIMARY KEY (\`id\`),
-			  KEY \`type\` (\`type\`),
-			  KEY \`user_id\` (\`user_id\`),
-			  KEY \`created_at\` (\`created_at\`)
-			) ENGINE=InnoDB DEFAULT CHARSET=latin1;`,
-			[savedDatabase || constants.saveQueryResultDb, constants.saveQueryResultTable],
-			conn,
-		);
-
-		if (!savedDatabase ) {
-
-			await this.mysql.query(
-				`update tb_credentials set db = ? where id = ?`,
-				[savedDatabase || constants.saveQueryResultDb, conn],
-				"write"
-			);
-		}
-
-		return savedDatabase || constants.saveQueryResultDb;
-	}
-
-}
-
-class OAuthConnection {
-
-	constructor(endpoint, provider) {
-
-		this.endpoint = endpoint;
-
-		Object.assign(this, provider);
-
-		this.provider = provider;
-
-		this.endpoint.assert(this.connection_id, 'Connection ID not found!');
-		this.endpoint.assert(this.client_id, 'Client ID not found!');
-		this.endpoint.assert(this.client_secret, 'Client secret not found!');
-	}
-
-	async validate() {
-
-		const
-			connection = {};
-
-		await this.getRefreshToken(connection);
-		await this.getAccessToken(connection);
-
-		this.endpoint.mysql.query(
-			'UPDATE tb_oauth_connections SET ? WHERE id = ?',
-			[connection, this.connection_id]
-		);
-	}
-}
-
-class GoogleAPIs extends OAuthConnection {
-
-	async getRefreshToken(connection) {
-
-		const
-			parameters = new URLSearchParams(),
-			options = {
-				method: 'POST',
-				body: parameters,
-				headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
-			};
-
-		parameters.set('code', this.endpoint.request.body.code);
-		parameters.set('client_id', this.client_id);
-		parameters.set('client_secret', this.client_secret);
-		parameters.set('redirect_uri', `http://${this.endpoint.request.headers.host}/connections-manager`);
-		parameters.set('grant_type', 'authorization_code');
-
-		let response = await fetch('https://www.googleapis.com/oauth2/v4/token', options);
-
-		response = await response.json();
-
-		this.endpoint.assert(response.refresh_token, `Refresh token not recieved from google, ${JSON.stringify(response)}!`);
-
-		this.refreshToken = connection.refresh_token = response.refresh_token;
-	}
-
-	async getAccessToken(connection) {
-
-		const
-			parameters = new URLSearchParams(),
-			options = {
-				method: 'POST',
-				body: parameters,
-				headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
-			};
-
-		parameters.set('refresh_token', connection.refresh_token);
-		parameters.set('client_id', this.provider.client_id);
-		parameters.set('client_secret', this.provider.client_secret);
-		parameters.set('grant_type', 'refresh_token');
-
-		let response = await fetch('https://www.googleapis.com/oauth2/v4/token', options);
-
-		response = await response.json();
-
-		this.endpoint.assert(response.access_token, 'Access Token not recieved form Google!');
-
-		this.accessToken = connection.access_token = response.access_token;
-		connection.expires_at = new Date(Date.now() + (response.expires_in * 1000)).toISOString().replace('T', ' ').replace('Z', '');
-	}
-
-	async test() {
-
-		this.endpoint.assert(this.access_token, 'Access Token not found!');
-		this.endpoint.assert(this.refresh_token, 'Refresh Token not found!');
-
-		await this.getAccessToken(this);
-
-		return this.access_token;
-	}
-}
-
-class FacebookMarketing extends OAuthConnection {
-
-	async getRefreshToken() {
-
-		return null;
-	}
-
-	async getAccessToken(connection) {
-
-		const
-			parameters = new URLSearchParams(),
-			options = {
-				method: 'POST',
-				body: parameters,
-				headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
-			};
-
-		parameters.set('client_id', this.provider.client_id);
-		parameters.set('redirect_uri', `http://${this.endpoint.request.headers.host}/connections-manager`);
-		parameters.set('client_secret', this.provider.client_secret);
-		parameters.set('code', this.endpoint.request.body.code);
-
-		let response = await fetch('https://graph.facebook.com/v3.0/oauth/access_token', options);
-
-		response = await response.json();
-
-		this.endpoint.assert(response.access_token, `Access Token not recieved form Facebook, ${JSON.stringify(response)}!`);
-
-		connection.access_token = response.access_token;
-		connection.expires_at = new Date(Date.now() + (response.expires_in * 1000)).toISOString();
-	}
-
-	async test() {
-
-		this.endpoint.assert(this.provider.access_token, 'Access Token not found!');
-
-		let  appAccessToken;
-
-		{
-			const
-				parameters = new URLSearchParams(),
-				options = {
-					method: 'POST',
-					body: parameters,
-					headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
-				};
-
-			parameters.set('client_id', this.provider.client_id);
-			parameters.set('client_secret', this.provider.client_secret);
-			parameters.set('grant_type', 'client_credentials');
-
-			let response = await fetch('https://graph.facebook.com/v3.0/oauth/access_token', options);
-
-			response = await response.json();
-
-			this.endpoint.assert(response.access_token, 'App Access Token not recieved form Facebook!');
-
-			appAccessToken = response.access_token;
-		}
-
-		{
-			const
-				parameters = new URLSearchParams(),
-				options = {
-					method: 'GET',
-					headers: { 'Content-Type': 'application/x-www-form-urlencoded' },
-				};
-
-			parameters.set('input_token', this.provider.access_token);
-			parameters.set('access_token', appAccessToken);
-
-			let response = await fetch('https://graph.facebook.com/v3.0/debug_token?' + parameters , options);
-
-			response = await response.json();
-
-			this.endpoint.assert(response.data && response.data.is_valid, 'Access token in not valid!');
-		}
-
-		return this.provider.access_token;
 	}
 }
